// src/types/draft-class.ts
export enum PlayerArchetype {
  Invalid_ = 0,
  QB_FieldGeneral = 1,
  QB_StrongArm = 2,
  QB_Scrambler = 3,
  HB_PowerBack = 4,
  HB_SpeedBack = 5,
  HB_ReceivingBack = 6,
  HB_BlockingBack = 7,
  FB_Blocking = 8,
  FB_Receiving = 9,
  WR_DeepThreat = 10,
  WR_Possession = 11,
  WR_Slot = 12,
  WR_Blocking = 13,
  TE_Receiving = 14,
  TE_Blocking = 15,
  TE_Balanced = 16,
  LT_PassBlocking = 17,
  LT_RunBlocking = 18,
  LG_PassBlocking = 19,
  LG_RunBlocking = 20,
  C_Balanced = 21,
  RG_PassBlocking = 22,
  RG_RunBlocking = 23,
  RT_PassBlocking = 24,
  RT_RunBlocking = 25,
  LE_PassRush = 26,
  LE_RunStuffing = 27,
  RE_PassRush = 28,
  RE_RunStuffing = 29,
  DT_PassRush = 30,
  DT_RunStuffing = 31,
  LOLB_PassRush = 32,
  LOLB_RunStuffing = 33,
  MLB_PassRush = 34,
  MLB_RunStuffing = 35,
  ROLB_PassRush = 36,
  ROLB_RunStuffing = 37,
  CB_ManToMan = 38,
  CB_ZoneCoverage = 39,
  CB_PressCoverage = 40,
  FS_DeepZone = 41,
  FS_ManToMan = 42,
  SS_RunSupport = 43,
  SS_ManToMan = 44,
  K_Accurate = 45,
  K_Power = 46,
  P_Power = 47,
  P_Accurate = 48,
}

export interface Prospect {
  // Personal Info
  id: string;
  firstName: string;
  lastName: string;
  homeState: number;
  homeTown: string;
  college: number;
  birthDate: number;
  age: number;
  heightInches: number;
  weight: number;
  position: number;
  archetype: PlayerArchetype;
  jerseyNum: number;
  assetName: string;

  // Draft Info
  draftable: number;
  draftPick: number;
  draftRound: number;

  // Ratings
  overall: number;
  acceleration: number;
  agility: number;
  awareness: number;
  ballCarrierVision: number;
  blockShedding: number;
  breakSack: number;
  breakTackle: number;
  carrying: number;
  catching: number;
  catchInTraffic: number;
  changeOfDirection: number;
  finesseMoves: number;
  hitPower: number;
  impactBlocking: number;
  injury: number;
  jukeMove: number;
  jumping: number;
  kickAccuracy: number;
  kickPower: number;
  kickReturn: number;
  leadBlock: number;
  manCoverage: number;
  passBlockFinesse: number;
  passBlockPower: number;
  passBlock: number;
  personality: number;
  playAction: number;
  playRecognition: number;
  powerMoves: number;
  pressCoverage: number;
  pursuit: number;
  release: number;
  shortRouteRunning: number;
  mediumRouteRunning: number;
  deepRouteRunning: number;
  runBlockFinesse: number;
  runBlockPower: number;
  runBlock: number;
  runningStyle: number;
  spectacularCatch: number;
  speed: number;
  spinMove: number;
  stamina: number;
  stiffArm: number;
  strength: number;
  tackle: number;
  throwAccuracyDeep: number;
  throwAccuracyMid: number;
  throwAccuracy: number;
  throwAccuracyShort: number;
  throwOnTheRun: number;
  throwPower: number;
  throwUnderPressure: number;
  toughness: number;
  trucking: number;
  zoneCoverage: number;
  morale: number;

  // Traits
  traitBigHitter: number;
  traitPossessionCatch: number;
  traitClutch: number;
  traitCoverBall: number;
  traitDeepBall: number;
  traitDlBullRush: number;
  traitDlSpinMove: number;
  traitDlSwimMove: number;
  traitDropsOpen: number;
  traitSidelineCatch: number;
  traitFightForYards: number;
  traitUnk1: number;
  traitHighMotor: number;
  traitAggressiveCatch: number;
  traitPenalty: number;
  traitPlayBall: number;
  traitPumpFake: number;
  traitLbStyle: number;
  traitSensePressure: number;
  traitUnk2: number;
  traitStripBall: number;
  traitTackleLow: number;
  traitThrowAway: number;
  traitTightSpiral: number;
  traitTendency: number;
  traitRunAfterCatch: number;
  devTrait: number;
  traitPredictability: number;

  // Appearance
  visuals: string;
  genericHead: GenericHead;
  handedness: number;
  portraitId: number;
  qbStyle: number;
  qbStance: number;
  unk3: number;
  unk4: number;
  unk5: number;
  unk6: number;
  visMoveType: number;
  unk8: number;
  commentaryId: number;
  unkByte2: number;
}

export interface DraftClass {
  year: number;
  prospects: Prospect[];
}

export enum Position {
  QB = 0,
  HB = 1,
  FB = 2,
  WR = 3,
  TE = 4,
  LT = 5,
  LG = 6,
  C = 7,
  RG = 8,
  RT = 9,
  LE = 10,
  RE = 11,
  DT = 12,
  LOLB = 13,
  MLB = 14,
  ROLB = 15,
  CB = 16,
  FS = 17,
  SS = 18,
  K = 19,
  P = 20,
}

export enum State {
  Alabama = 0,
  Alaska = 1,
  Arizona = 2,
  Arkansas = 3,
  California = 4,
  Colorado = 5,
  Connecticut = 6,
  Delaware = 7,
  Florida = 8,
  Georgia = 9,
  Hawaii = 10,
  Idaho = 11,
  Illinois = 12,
  Indiana = 13,
  Iowa = 14,
  Kansas = 15,
  Kentucky = 16,
  Louisiana = 17,
  Maine = 18,
  Maryland = 19,
  Massachusetts = 20,
  Michigan = 21,
  Minnesota = 22,
  Mississippi = 23,
  Missouri = 24,
  Montana = 25,
  Nebraska = 26,
  Nevada = 27,
  NewHampshire = 28,
  NewJersey = 29,
  NewMexico = 30,
  NewYork = 31,
  NorthCarolina = 32,
  NorthDakota = 33,
  Ohio = 34,
  Oklahoma = 35,
  Oregon = 36,
  Pennsylvania = 37,
  RhodeIsland = 38,
  SouthCarolina = 39,
  SouthDakota = 40,
  Tennessee = 41,
  Texas = 42,
  Utah = 43,
  Vermont = 44,
  Virginia = 45,
  Washington = 46,
  WestVirginia = 47,
  Wisconsin = 48,
  Wyoming = 49,
  NonUS = 50,
  CanadaAlberta = 51,
  CanadaBritishColumbia = 52,
  CanadaOntario = 53,
  CanadaQuebec = 54,
  AmericanSamoa = 55,
  AustraliaNewSouthWales = 56,
  AustraliaVictoria = 57,
  Austria = 58,
  Bahamas = 59,
  Belgium = 60,
  Belize = 61,
  Brazil = 62,
  Cameroon = 63,
  England = 64,
  France = 65,
  Germany = 66,
  Ireland = 67,
  Jamaica = 68,
  Japan = 69,
  Mexico = 70,
  Nigeria = 71,
  PuertoRico = 72,
  Scotland = 73,
  SouthAfrica = 74,
  Spain = 75,
  Turkey = 76,
  Argentina = 77,
  UnitedArabEmirates = 78,
  Italy = 79,
  INVALID = 80,
}

export enum Handedness {
  Right = 0,
  Left = 1,
}